Fixed: Improvements to the emscripten build script
Previously the build script would fail with a python error if the emscripten compiler was not found. This commit adds support for checking if the compiler exists on the path, and failing with an error message if it does not.

Also fixed the shebang to use /usr/bin/env bash in case bash was not at /bin/bash.