WIP: accidentals drawn in clusters
Chord::GenerateAccidList -> ResetAccidList, m_accidList cached as a member of chord.

Chord::ResetAccidSpace resets the accidental space, takes whatever the current fullUnit is as a parameter.

Added "save" parameter to CalculateAccidX that toggles whether or not it writes the data to the x-position of the accidental or to the accidSpace.

Moved above code out of view_element, added an initial pass through all notes to determine accidental overlaps before drawing.