more SVG clean up
removes more unnecessary initial values:
* fill=black" and fill-rule="nonzero"
for rect rx gets only set if it’s non zero:
> If neither ‘rx’ nor ‘ry’ are properly specified, then set both rx and
ry to 0.

And finally every path now gets a proper stroke. This seems to blow up
the SVG, but setting black in stroke as a default value gives every sub
element an outline, which is bad SVG and so we would have to set
stroke-width to zero for every non-path element. Rendering text gets so
cleaner and coloring text gets easier as we only have to set fill.