Keep pitch element position with 2nd clef insert
When a 2nd (or 3rd or 4th etc.) clef is inserted in a layer, keep the
position of elements that have their position defined
relative to that clef the same. This is done by shifting their pitch by
twice the difference in line numbers between the new clef and the
previous one.

See DDMAL/Neon2#273.