fix(tutorial 3): fix broken css for tutorial 3
According to this: https://stackoverflow.com/a/5546296, the selector .className:nth-child(2) does not really select the second child with className but the second child of the element type. So it will match always the second g element in this case as there are only g elements inside .measure instead of matching the second g element with class name staff. Changing the selector to nth-of-type() has the same effect. Using these selectors, you can only match the nth child of a given element.

Apparently, there seems to be no similar CSS Selector to match e.g. the second (or nth) child of a given class.

This fix makes use of the solution mentioned here: https://stackoverflow.com/a/6447072 . It selects all siblings of the nth child, applies the attributes to all of them and then removes them from all n+1 siblings.