`@skew` to `@rotate`
Change skew to the TEI `@rotate` attribute. This is kind of a messy switch
to get something working and should be redone if/when `@rotate` is
properly in MEI.

Important to note that TEI `@rotate` goes CW and does NOT follow the right hand
rule. That's been implemented here for consistency. I apologize to my
high school physics teacher.