Avoid crash on empty beams. Fixes fix #1765
beams with `@sameas` are expected to be empty, so not need to check on this here. Empty beams without `@sameas` are not valid MEI. However, they can still be loaded into Verovio, especially if the validation does not support Schematron. For this reason, it is good to check that the beam as not empty here. Test is an OR and not AND.