Render custos y-position from pitch info
Need to do this to ensure a match between position of the custos on the
staff and pitch info since small errors in facsimile y-position can
compound through editing and lead to big problems.

Should fix issues with facsimile in Neon due to correction of my own
mistake in #1979. (Sorry about that!)