Do not consider order when assembling syl bb
Previously, it was assumed that the first bounding box would be the
leftmost. This is what *should* happen, but is not actually a
requirement. This can lead to issues where a rightmost edge is to the
left of the leftmost edge (see DDMAL/Neon#656). This commit removes this
assumption and fixes the referenced issue.