New: Generate Utility for working with font files
This commit adds a new `generate.py` script for working with the font files in Verovio.

It is meant to consolidate similar functions in the older scripts, which have been moved to the `old-code` directory for reference.
(These may be deleted later once we are happy that the new script works as expected.)

Calling the `generate.py` script with a `--help` argument will print the sub-commands. Calling `generate.py help [sub-command]` will
provide more details on how the sub-command works.

It requires the `svgpathtools` Python module, and the `fontforge` binary (if you want to create woff2 and svg fonts). It no longer requires
`lxml`, and uses the built-in Python ElementTree module. (The attribute ordering problem was fixed in Python 3.8).

This script has been tested against the existing output for all the fonts, and has been verified to produce the same output. (The biggest difference
is that, since we are using the `ElementTree` writer directly, there is a space between the last attribute and the `/>` at the end of the tag).

The Gootville font was also renamed from `Gootville-1.2.otf` to `Gootville.otf` to be consistent with the naming scheme of the other fonts.

A separate commit will hold the re-generated files so that they can be compared for changes.