Fixed: Problematic font attributes
Previously there was a bug in the bounding boxes python script, where a value of "0.0" was used as the guard value in an if statement. Since 0.0 evaluates to a falsey value, the "w" value was overwritten by the h-a-x value.

When a new method of checking the value could be cast to a float was introduced, this fixed the bug, but it means that it changes the "w" values in some glyphs.