Fixed: Control timestamp of font
Fontforge automatically adds the current date and time to every font file, which means
that rebuilding the files generates a new base64 value.

This commit adds the "SOURCE_DATE_EPOCH" environment variable which FF uses to
set the value directly. This is set to the last modified time of the source font file, so if that
changes, then the output from fontforge should also change. Otherwise, regenerating the
font should produce the same encoding.