Fixed: Add a command for checking a font
The `check [fontname]` command will check a font's full SVG file against the list
of supported glyphs, and will report the code of any glyph that is supported by
Verovio, but that is not in that font.

Optionally, it can also show the list of glyphs in the font that are not supported
by Verovio by passing the `--show-unsupported` flag.

Fixes #3035