WIP: Recognize CMAKE_INSTALL_PREFIX
By setting an absolute path to the install directory, the cmake build was ignoring the CMAKE_INSTALL_PREFIX

By changing it to a relative path the CMAKE_INSTALL_PREFIX seems to work as expected.

This also alters the path to the resource directory for Verovio to be configurable at compile-time.
This is hard-coded in the source, so it uses a command-line define, `VEROVIO_RESOURCE_DIRECTORY`,
to configure this. By default, it will use the CMAKE_INSTALL_PREFIX/share/verovio

These changes make it possible to adapt the installation of Verovio to different OS filesystem layouts. (e.g., homebrew on Mac uses
/opt/homebrew instead of /usr/local now)