New: Enable "fastdeps" in SWIG
This option generates a Python binding that executes the proxied C++ code faster.

Adding the "olddeps" option generates a Python proxy with the correct method signatures so that calling code can treat it like Python, but the actual call is done with the proxied property.

Note that another option, "builtin" was also tried, but this broke existing implementations so it was not suitable.

For a comparison of the different methods see: https://www.swig.org/Doc4.2/Python.html#Python_optimization